refactoring and change a bug in editPage
