+++ clientSide/src/components/EditBooking/EditBooking.jsx
@@ -8,55 +8,13 @@ import { onChangeHandler } from "../../functions/onChangeHandlers";
 import DropDown from "../DropDown/DropDown";
 import Button from "../Button/Button";
 import NavBar from "../NavBar/Navbar";
-import Parse from "parse";
-import {
-  getOfficePointer,
-  getBookingStatePointer,
-  getCarGroupPointer,
-} from "../../parse-functions/bookingComponentFunctions";
 import PageTitle from "../PageTitle/PageTitle";
+import { updateBooking } from "../../parse-functions/updateFunctions";
+
 const EditBooking = () => {
   const { bookingId } = useParams();
   const navigate = useNavigate();
   const [formData, setFormData] = useState([]);
-  console.log(bookingId);
-  console.log(formData);
-
-  const updateBooking = async (e, bookingId, formData, navigate) => {
-    e.preventDefault();
-    const pickupOfficePointer = await getOfficePointer(formData.pickUpOffice);
-    const returnOfficePointer = await getOfficePointer(formData.returnOffice);
-    const carGroupPointer = await getCarGroupPointer(formData.carGroup);
-    const bookingStatePointer = await getBookingStatePointer(
-      formData.bookingState
-    );
-    console.log("pickuppointer", pickupOfficePointer);
-    console.log("returnPointer", returnOfficePointer);
-    console.log("bookingStatePointer", bookingStatePointer);
-    console.log("carGroupPointer", carGroupPointer);
-    const Booking = Parse.Object.extend("Booking");
-    const query = new Parse.Query(Booking);
-
-    const object = await query.get(bookingId);
-    object.set("pickUpTime", new Date(formData.pickUpDate));
-    object.set("carGroup", carGroupPointer);
-    object.set("returnOffice", returnOfficePointer);
-    object.set("pickUpOffice", pickupOfficePointer);
-    object.set("bookingState", bookingStatePointer);
-    object.set("returnTime", new Date(formData.returnDate));
-    console.log("bookingObject", object);
-
-    try {
-      const response = await object.save();
-      console.log("Booking updated", response);
-      navigate("/find-booking");
-    } catch (error) {
-      console.error("Error while updating Booking", error);
-    }
-    //   catch (error) {
-    //   console.error("Error while retrieving object Booking", error);
-    //   }
-  };
 
   return (
     <>
